+++ backend/src/docpool.theme/docpool/theme/resources/default.js
@@ -61,17 +61,31 @@ jQuery(document).on("click", ".cat_filter #reset_filter", function () {
 
 $(function () {
   // Adds the docType string to the edit heading
+  // Only for Admins
   // https://redmine-koala.bfs.de/issues/5282
-  var type = $("#form-widgets-docType");
-  if (type[0] != null && type[0].selectedIndex != null) {
-    var dtParam = getURLParam("form.widgets.docType:list");
-    if (typeof dtParam != "undefined") {
-      type.val(dtParam);
+  // https://redmine-koala.bfs.de/issues/6057
+  if (isAdminOrContentAdmin == null) {
+    var type = $("#form-widgets-docType");
+    if (type?.[0]) {
+      var value = null;
+
+      // For Add-Form as it has display mode = "input"
+      if (type[0].selectedIndex != null) {
+        const docTypeFormFieldList = type[0].options[type[0].selectedIndex];
+        value = docTypeFormFieldList.text;
+        $(docTypeFormFieldList).parent().parent().hide();
+      }
+
+      // For Edit-Form as it has display mode = "display"
+      if (type[0].querySelector(".selected-option")) {
+        const docTypeFormField = type?.[0]?.querySelector(".selected-option");
+        value = docTypeFormField.textContent;
+        // Remove the FormField
+        $(docTypeFormField).parent().parent().hide();
+      }
+      var title = $("h1.documentFirstHeading").text();
+      $("h1.documentFirstHeading").text(value ? `${title} (${value})` : title);
     }
-    var option = type[0].options[type[0].selectedIndex];
-    var value = option.text;
-    var title = $("h1.documentFirstHeading").text();
-    $("h1.documentFirstHeading").text(title + " (" + value + ")");
   }
 });